requirements.txt
# Core Dependencies
streamlit==1.51.0
python-dotenv==1.0.1

# OpenAI Realtime API
websocket-client==1.8.0

# Audio Processing
PyAudio==0.2.14

# Image Processing
Pillow==11.0.0
